Added TAILQ for sent delay_req
In a ptp unaware network (like the telecom profile for frequency sync G.8265.1), both the RTD and the PDV can be substantially higher than in a ptp aware network. To achieve more accurate measurements, the rate may need to be configured higher to get more data and increase the chance of lucky packets.
In a combination of a high configured rate of delay_req and high RTD/PDV in network, the risk that the response from the previously sent delay_req have not been received before a new delay_req is sent also become high. In that case, the need of storing more than the latest sent delay_req arise.

Signed-off-by: Anders Selhammer <anders.selhammer@est.tech>