phc2sys: Postpone adding of servo to clock.
Instead of unconditionally creating a servo for each clock in their
initialization, add the servo later on the first update of the clock,
when it is known the clock needs to be synchronized.

This fixes an issue with phc2sys disrupting the system clock when
it is synchronized by another process and should be used by phc2sys only
as a source. Creating a servo for the system clock caused a reset of its
frequency and status.

Reported-by: Russell King <linux@armlinux.org.uk>
Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>