ts2phc: nmea: Add a configuration option for the current leap seconds file.
The PTP time scale is TAI, and a PTP GM needs to provide it.  However, most
GPS devices only provide UTC time of day information, and they do not, in
general, offer any kind of reliable, standardized leap seconds data.  After
all, this information is only broadcast every 12.5 minutes through the GPS
network, and that is far too long for a PTP GM to wait after startup.

This patch allows the built in leap seconds table to be overridden by an
up to date file provided in the environment.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>