ts2phc: nmea: Drop time of day readings older than five seconds.
The nmea code estimates the device's current time by adding the duration
from the last reading as measured with the local montonic clock.  This
method is fine for a short duration, but the likely frequency offset
between the two clocks limits the long term accuracy.  In addition, if no
recent RMC record is available, probably some kind of error has occurred.
This patch limits the age of the RMC record used for ToD to five seconds.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>