phc2sys: Fix null pointer de-reference in manual mode.
If both the -w and -O command line options are specified (or when
using -w when both source and destination clocks are PHCs), then
pointer to the PMC agent will be incorrectly freed.

Fix the segfault by introducing a method to "disable" the agent as was
done before the PMC agent code was introduced.

Unfortunately the resulting PMC agent API now has both create/destroy
and init/disable methods.  This clunky arrangement can be cleaned up
later on, but it entails re-factoring the phc2sys program even more.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>
Fixes: 8266987 ("pmc_agent: Hide the implementation.")