phc2sys: Replace magical test with a proper test.
When using a PPS time source with the -w flag, the program closes the PMC
agent's connection once the ptp4l service has appeared and the UTC offset
has been optionally queried.

In order to tell whether PPS is the time source, the code tests for

    src->clkid == CLOCK_INVALID

which is completely opaque and non-obvious.  Presently only the case where
src_name is NULL (and pps_fd is valid) will trigger this test, but that is
only incidental.

Clarify the intention of the code by using a proper test.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>