phc2sys: Replace yet another magical test with a proper test.
In the manual wait mode, if the system clock is not being used, the
program closes the PMC agent's connection once the ptp4l service has
appeared.  Presumably this is done because the TAI-UTC offset plays no
role in synchronizing two PHC devices.

In order to tell whether the system clock is being used, the code tests for

    src->clkid != CLOCK_REALTIME && dst->clkid != CLOCK_REALTIME

which is fairly non-obvious.  Clarify the intention of the code by
using a proper test.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>