phc2sys: Fix regression in the automatic mode.
Commit ac7d69bbc476 ("pmc_agent: Convert the method that queries the
port properties.") had the well meant intention of the cleaning up the
error code semantics of the port properties query function.  However,
that commit mixed up the normal, external semantics of zero meaning
success with the internal semantics where zero is an error.  Correct
the issue by replacing the hard coded number with the proper macro.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>
Fixes: ac7d69bbc476 ("pmc_agent: Convert the method that queries the port properties.")