phc2sys: Ensure PHC source when using PPS mode.
Using the mode PPS without a PHC source causes a segfault in strdup()
in the clock_add() function.  This bug appeared between version 1.4
and 1.5 of the stack.  However, even before 1.4 the PPS mode was
already broken in this respect.  Considering the lack of complaints
over the years, clearly nobody ever uses this mode.

In practice, the "clock compare" method has always provided better
synchronization performance than using the PPS mode.  Probably we
could remove it, and nobody would notice.  Keeping the mode, however,
is useful to demonstrate the superiority of other method.

Remove the code paths of non-working PPS variant.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>