port: Fix link down/up to continue using phc_index set from command line -p option.
In the scenario where a port link goes down and up, current code checks
the port's phc_index against the interface's phc_index and if they are
different will set the port phc_index to the interface phc_index.

If the phc_index was initially set by the command line -p option, then we
end up using the wrong phc_index.

Fix is to skip updating the port phc_index with the interface phc_index
when port link is back up if it was initialy set from the command line.

Signed-off-by: Vincent Cheng <vincent.cheng.xh@renesas.com>